Issue #5104: created internal test module package
