+++ src/test/resources/com/puppycrawl/tools/checkstyle/ant/checkstyleanttask/InputCheckstyleAntTaskConfigCustomCheckerRootModule.xml
@@ -3,5 +3,5 @@
         "-//Puppy Crawl//DTD Check Configuration 1.3//EN"
         "http://checkstyle.sourceforge.net/dtds/configuration_1_3.dtd">
 
-<module name="com.puppycrawl.tools.checkstyle.CheckerStub">
+<module name="com.puppycrawl.tools.checkstyle.internal.testmodules.CheckerStub">
 </module>
