+++ src/test/resources/com/puppycrawl/tools/checkstyle/main/InputMainConfig-custom-root-module.xml
@@ -3,6 +3,6 @@
           "-//Puppy Crawl//DTD Check Configuration 1.3//EN"
           "http://checkstyle.sourceforge.net/dtds/configuration_1_3.dtd">
 
-<module name="com.puppycrawl.tools.checkstyle.TestRootModuleChecker">
+<module name="com.puppycrawl.tools.checkstyle.internal.testmodules.TestRootModuleChecker">
     <property name="property" value="${lineLength.severity}" default=""/>
 </module>
