+++ src/test/resources/com/puppycrawl/tools/checkstyle/main/InputMainConfig-multi-thread-mode.xml
@@ -3,7 +3,7 @@
           "-//Puppy Crawl//DTD Check Configuration 1.3//EN"
           "http://checkstyle.sourceforge.net/dtds/configuration_1_3.dtd">
 
-<module name="com.puppycrawl.tools.checkstyle.TestRootModuleChecker">
+<module name="com.puppycrawl.tools.checkstyle.internal.testmodules.TestRootModuleChecker">
   <property name="property" value="test" default="test"/>
   <module name="Checker">
     <module name="TreeWalker">
